+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - Fix: Skyscraper no longer throws exceptions when using processed-cache
   and some of the processors don’t have `:cache-template`.
+- Fix: Skyscraper no longer throws exceptions when the server returns
+  multiple Content-Type headers.
 
 ## 0.3.1 (2022-07-31)